+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Linting
 
 * Adjusted linting to enable `patch` branches from being tested
+* `.travis.yml` now optional, `.circle.yml` and `Singularity` must not be present
 
 ### Template